feat: add session switching to the root action menu
