+++ app/actions/root.ts
@@ -3,6 +3,7 @@ import { rootDeveloperActions } from "./definitions/developer";
 import { rootDocumentActions } from "./definitions/documents";
 import { rootNavigationActions } from "./definitions/navigation";
 import { rootSettingsActions } from "./definitions/settings";
+import { rootTeamActions } from "./definitions/teams";
 import { rootUserActions } from "./definitions/users";
 
 export default [
@@ -12,4 +13,5 @@ export default [
   ...rootNavigationActions,
   ...rootSettingsActions,
   ...rootDeveloperActions,
+  ...rootTeamActions,
 ];
