+++ shared/i18n/locales/en_US/translation.json
@@ -45,6 +45,8 @@
   "Appearance": "Appearance",
   "Change theme": "Change theme",
   "Change theme to": "Change theme to",
+  "Change or create team": "Change or create team",
+  "Select a team": "Select a team",
   "Invite people": "Invite people",
   "Collection": "Collection",
   "Debug": "Debug",
@@ -294,7 +296,6 @@
   "New child document": "New child document",
   "New document in <em>{{ collectionName }}</em>": "New document in <em>{{ collectionName }}</em>",
   "New template": "New template",
-  "Switch team": "Switch team",
   "Link copied": "Link copied",
   "Revision options": "Revision options",
   "Restore version": "Restore version",
